Refactor: create UI components for permissions overview
